+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Released: Upcoming
 - Attention: The "or" validator now accepts complete rules instead of parts so the RulesBuilder and actually more
   complex rules can be used
 - Attention: The RulesBuilder function "getRules" got renamed to "build" with an additional reset of the rules builder
+- Attention: Renamed the RuleBuilder function "addRule" to "addFieldRule" and added a new version of "addRule" without a
+  $field parameter so Rules for collection, nested and or validators can be build with it
 - Added a collection validator
 - Added a nested validator